InsertionSort test and JavaUtil fixed
java util generateTimingReport method did not actually generate a timing report, but it does now. Also the insertionSort method is tested and works. The threshold method for MergeSort does in fact make it insertionSort, but we still have the problem where we run out of heap space because the method does infinite recursion.